refactor: 앨범 정보 type 정리 및 widgets/album 컴포넌트들 ui 폴더로 이동
