+++ client/src/widgets/albums/ui/AlbumList.tsx
@@ -1,13 +1,6 @@
 import LogoAlbum from '@/assets/logo-album-cover.png';
 import { AlbumCard } from './AlbumCard';
-
-export interface Album {
-  id: number;
-  title: string;
-  artist: string;
-  tags: string[];
-  coverImage: string;
-}
+import { Album } from '@/entities/album/types';
 
 const mockAlbums: Album[] = Array.from({ length: 7 }, (_, index) => ({
   id: index,
@@ -21,7 +14,7 @@ export function AlbumList() {
   return (
     <div className="text-grayscale-50">
       <p className="mt-[70px] mb-7 text-3xl font-bold">최근 등록된 앨범</p>
-      <ul className="flex flex-row gap-9 justify-between	">
+      <ul className="flex flex-row flex-wrap gap-9 ">
         {mockAlbums.map((album) => (
           <AlbumCard key={album.id} album={album} />
         ))}
